Move FES module out of core
