+++ src/friendly_errors/browser_errors.js
