+++ src/friendly_errors/stacktrace.js
